# peanut-suite/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Peanut Suite REST API
  description: |
    Complete REST API documentation for Peanut Suite WordPress plugin.

    ## Authentication

    The API supports two authentication methods:

    ### 1. WordPress Nonce (Cookie Auth)
    For browser-based requests, include the `X-WP-Nonce` header with a valid WordPress REST nonce.

    ### 2. API Key (Bearer Token)
    For programmatic access, use Bearer token authentication:
    ```
    Authorization: Bearer psk_your_api_key_here
    ```

    API keys can be scoped to specific resources:
    - `links:read`, `links:write`
    - `utms:read`, `utms:write`
    - `contacts:read`, `contacts:write`
    - `analytics:read`

    ## Rate Limiting

    - Login attempts: 5 per 5 minutes
    - Team invites: 10 per hour
    - General API: 60 requests per minute

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`
    - `X-RateLimit-Remaining`
    - `X-RateLimit-Reset`

  version: 4.2.26
  contact:
    name: Peanut Suite Support
    url: https://peanutgraphic.com
  license:
    name: Proprietary
servers:
  - url: /wp-json/peanut/v1
    description: WordPress REST API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API Key authentication
    nonceAuth:
      type: apiKey
      in: header
      name: X-WP-Nonce
      description: WordPress REST nonce
...
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: Get plugin settings
      security:
        - nonceAuth: []
      responses:
        '200':
          description: Plugin settings

    post:
      tags: [Settings]
      summary: Update plugin settings
      security:
        - nonceAuth: []
      responses:
        '200':
          description: Settings updated
  /settings/export:
    get:
      tags: [Settings]
      summary: Export all settings
      security:
        - nonceAuth: []
      responses:
        '200':
          description: Exported settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename:
                    type: string
                  data:
                    type: object
  /settings/import:
    post:
      tags: [Settings]
      summary: Import settings
      security:
        - nonceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
      responses:
        '200':
          description: Settings imported
  /settings/clear-cache:
    post:
      tags: [Settings]
      summary: Clear all caches
      security:
        - nonceAuth: []
      responses:
        '200':
          description: Cache cleared
  /settings/delete-all:
    delete:
      tags: [Settings]
      summary: Delete all plugin data
      security:
        - nonceAuth: []
      parameters:
        - name: confirm
          in: query
          required: true
          schema:
            type: string
            enum: [DELETE_ALL_DATA]
      responses:
        '200':
          description: All data deleted

  # ==================== DASHBOARD ====================
...